backend/requirements.txt
fastapi
uvicorn
requests
pulp
supabase
python-dotenv
